renamed test file name to test_pre_processing added extra test
